requirements.txt
annotated-types==0.7.0
anyio==4.10.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
distro==1.9.0
Flask==3.1.2
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
oauthlib==3.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic_core==2.33.2
pyparsing==3.2.4
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
runwayml==3.14.1
sniffio==1.3.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
